refactor(base): colocate search-related tests
